requirements.txt
argparse
datetime
uuid
cloudpickle==1.3.0
cython==0.29.20
ipdb==0.13.13
matplotlib==3.2.2
numpy==1.18.5
pandas==1.0.5
pickleshare==0.7.5
pillow==7.1.2
scipy==1.5.0
seaborn==0.12.2
torch==1.5.1
torchvision==0.6.1
tqdm==4.46.1
urllib3==1.25.9
